requirements.txt
tensorflow
numpy
Pillow
mlflow
gymnasium
scikit-learn
joblib
Janome
